module_3/requirements.txt
blinker==1.9.0
click==8.2.1
flask==3.1.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
pip==25.1.1
psycopg2==2.9.10
python-dotenv==1.1.0
werkzeug==3.1.3
pylint==3.3.7
pytest==8.3.5
yapf==0.43.0